+++ apps/backend/src/application/use-cases/friend/index.ts
@@ -1,5 +1,5 @@
-export { CreateFriendInviteTokenUseCase } from "./create-friend-invite-token";
-export { ApplyFriendUseCase } from "./apply-friend";
+export { CreateFriendInviteTokenUseCase } from "./create-friend-invite-token.usecase";
+export { ApplyFriendUseCase } from "./apply-friend.usecase";
 
-export type { ICreateFriendInviteTokenUseCase } from "./interfaces/create-friend-invite-token.interface";
-export type { IApplyFriendUseCase } from "./interfaces/apply-friend.interface";
+export type { ICreateFriendInviteTokenUseCase } from "./interfaces/create-friend-invite-token.usecase.interface";
+export type { IApplyFriendUseCase } from "./interfaces/apply-friend.usecase.interface";
