+++ apps/backend/src/application/use-cases/friend/interfaces/apply-friend.usecase.interface.ts
