+++ apps/backend/src/application/use-cases/friend/interfaces/create-friend-invite-token.usecase.interface.ts
